requirements-dev.txt
-r requirements.txt
pytest>=8,<9
pytest-django>=4.8,<5
ruff>=0.8,<1
black>=24,<26
